move shared structs out of client and into nomad
